Initial commit. Very hackish way to see the CI status.
This is probably the dirtiest hackish way to show the Jenkins wall display in an Android device.
It works with the Github Auth plugin for Jenkins (only tested with Novoda's server).
Needs to be cleaned up, and polished A LOT! (rotation changes, errors, harcoded strings & dimensions, layouts, etc.) but at least works.